refactor: extract IPointerEventsController to single file
- also moved PointerEventsController.cs to its own folder (it was alone in Controllers root folder).
