+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PointerEventsController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 779b6e87d4fc3ec4ea4eaa86f27c197b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
