+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PointerEventsController/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a5edbad1a7d40a442852b83317bec281
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
