+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PointerEventsController/Interfaces/IPointerEventsController.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace DCL
+{
+    public interface IPointerEventsController
+    {
+        void Initialize();
+        void Update();
+        void Cleanup();
+        Ray GetRayFromCamera();
+    }
+}
