+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PointerEventsController/Interfaces/IPointerEventsController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8c9f360f165f4a3ebfe2bacd3f8bd89e
+timeCreated: 1620782817
